#162 feat: 선호 여행지 기반 여행지 필터링 전략패턴 구현
