+++ backend/src/main/java/moheng/recommendtrip/application/RecommendTripService.java
@@ -5,8 +5,8 @@ import moheng.member.domain.Member;
 import moheng.member.domain.repository.MemberRepository;
 import moheng.member.exception.NoExistMemberException;
 import moheng.recommendtrip.domain.RecommendTrip;
-import moheng.recommendtrip.domain.TripFilterStrategy;
-import moheng.recommendtrip.domain.TripFilterStrategyProvider;
+import moheng.recommendtrip.domain.tripfilterstrategy.TripFilterStrategy;
+import moheng.recommendtrip.domain.tripfilterstrategy.TripFilterStrategyProvider;
 import moheng.recommendtrip.domain.RecommendTripRepository;
 import moheng.recommendtrip.dto.RecommendTripCreateRequest;
 import moheng.recommendtrip.exception.LackOfRecommendTripException;
@@ -47,34 +47,9 @@ public class RecommendTripService {
     }
 
     public FindTripsResponse findRecommendTripsByModel(final long memberId) {
-        final Member member = memberRepository.findById(memberId)
-                .orElseThrow(NoExistMemberException::new);
-        final List<MemberTrip> memberTrips = memberTripRepository.findByMember(member);
-        final List<RecommendTrip> recommendTrips = recommendTripRepository.findTop10ByMember(member);
-        final Map<Long, Long> preferredLocations = findMemberPreferredLocations(memberTrips, recommendTrips);
-        final TripFilterStrategy tripFilterStrategy = tripFilterStrategyProvider.findFilterTripsByFilterStrategy(PREFERRED_LOCATIONS_STRATEGY);
-        final List<Trip> filteredTrips = tripFilterStrategy.execute();
+        final TripFilterStrategy tripFilterStrategy = tripFilterStrategyProvider.findTripsByFilterStrategy(PREFERRED_LOCATIONS_STRATEGY);
+        final List<Trip> filteredTrips = tripFilterStrategy.execute(memberId);
         return new FindTripsResponse(tripKeywordRepository.findByTrips(filteredTrips));
-    }
-
-    private Map<Long, Long> findMemberPreferredLocations(final List<MemberTrip> memberTrips, final List<RecommendTrip> recommendTrips) {
-        validateRecommendTrips(recommendTrips);
-        final Map<Long, Long> preferredLocations = new HashMap<>();
-        for (final RecommendTrip recommendTrip : recommendTrips) {
-            final Trip trip = recommendTrip.getTrip();
-            final MemberTrip memberTrip = memberTrips.stream()
-                    .filter(mt -> mt.getTrip().getContentId().equals(trip.getContentId()))
-                    .findFirst()
-                    .orElseThrow(() -> new NoExistMemberTripException("존재하지 않는 멤버의 선호 여행지입니다."));
-            preferredLocations.put(trip.getContentId(), memberTrip.getVisitedCount());
-        }
-        return preferredLocations;
-    }
-
-    private void validateRecommendTrips(final List<RecommendTrip> recommendTrips) {
-        if(recommendTrips.size() < MIN_RECOMMEND_TRIPS_COUNT) {
-            throw new LackOfRecommendTripException("추천을 받기위한 선호 여행지 데이터 수가 부족합니다.");
-        }
     }
 
     @Transactional
