--- backend/src/main/java/moheng/recommendtrip/domain/TripFilter.java
@@ -1,4 +1,0 @@
-package moheng.recommendtrip.domain;
-
-public class TripFilter {
-}
