--- backend/src/main/java/moheng/recommendtrip/domain/TripFilterByLiveInfoManager.java
@@ -1,4 +1,0 @@
-package moheng.recommendtrip.domain;
-
-public class TripFilterByLiveInfoManager {
-}
