+++ backend/src/main/java/moheng/recommendtrip/domain/tripfilterstrategy/TripFilterStrategy.java
@@ -1,4 +1,4 @@
-package moheng.recommendtrip.domain;
+package moheng.recommendtrip.domain.tripfilterstrategy;
 
 import moheng.trip.domain.Trip;
 
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface TripFilterStrategy {
     boolean isMatch(final String strategyName);
-    List<Trip> execute();
+    List<Trip> execute(final long memberId);
 }
